// src/data/Types.ts
export interface BestMenu {
   image: string
   name: string
   price: number
   star1: string
   star2: string
   star3: string
   star4: string
   star5: string
   bestM: any
   setBestM: any
   // onChange: (e:any) => {} | void
}

export interface BestChoice {
   image: string
   name: string
   price: number
   star1: string
   star2: string
   star3: string
   star4: string
   star5: string
   bestC: any
   setBestC: any
}

export interface MenuType {
   image: string
   name: string
   price: number
   star1: string
   star2: string
   star3: string
   star4: string
   star5: string
   typek: any
   setTypek: any
}

export interface MenuType2 {
   image: string
   name: string
   price: number
   star1: string
   star2: string
   star3: string
   star4: string
   star5: string
   typeq: any
   setTypeq: any
}

export interface MenuType3 {
   image: string
   name: string
   price: number
   star1: string
   star2: string
   star3: string
   star4: string
   star5: string
   typer: any
   setTyper: any
}

export interface CommentType {
   image: string
   comment: string
   star1: string
   star2: string
   star3: string
   star4: string
   star5: string
}